Rename build module to assembler to avoid Cargo build script confusion
`src/build.rs` shared its name with the Cargo build script convention
(a `build.rs` at the project root). While Cargo only treats a root-level
`build.rs` as a build script, the identical filename caused unnecessary
confusion for contributors unfamiliar with the codebase.

Rename the module to `assembler` to make its role clear: it assembles
the output `.gitignore` content from the parsed `.gitignore.in` script.

No behaviour change.
