+++ src/assembler.rs
